Keep power effects in time on long-running server worlds
Effects were timed from the world's game time plus the frame's partial
tick, written as ClientState.now()+partial or getGameTime()+getPartialTick().
Java adds a long and a float in float, which holds 24 bits. Once a world
has run 2^23 ticks (under five days of uptime) the partial tick rounds
away; past 2^25 (about nineteen days) the sum moves in four-tick steps,
and past 2^26 in eight. On a server that runs day and night, portals,
Time Branch, meteors, grips, wounds, the frost aim and the HUD meters
therefore started late and moved in jumps. A new single-player world
never gets there. The cape was unaffected because its cloth runs on
the wearer's own tickCount.

ClientState gains exact helpers, and every such site uses them:
- time(partial): the render clock in double.
- since(stamp, partial): ticks since a stamp, subtracted in whole ticks
  first.
- cycle(phase): a palette phase with whole turns dropped before it
  narrows to float. The palettes repeat every 1, so the colour is
  unchanged.
- wave(partial): a float clock folded every 2^20 ticks for purely
  cosmetic float maths: the Scepter's pulses and the Fracture screen.

The sun and Paradise sky spins now wrap at 360 degrees before narrowing.
On a new world every result is the same as before.

verifyRenderClock, hooked into check, fails the build if a renderer
adds a partial tick to the clock, or feeds a palette, as a float again.
It flags 59 sites in the previous tree and none now.

+++ src/main/java/com/hexgodofstories/client/VoidSeaWaveRenderer.java
@@ -96,7 +96,7 @@ public final class VoidSeaWaveRenderer {
         // them rather than swimming along with it.
         originX = Math.floor(camX / CELL) * CELL - radius;
         originZ = Math.floor(camZ / CELL) * CELL - radius;
-        double time = mc.level.getGameTime() + event.getPartialTick();
+        double time = mc.level.getGameTime() + (double) event.getPartialTick();
 
         // The dozen or so waves that can reach this patch, resolved once for the whole frame rather
         // than re-derived at each of several thousand grid points.
